Adjust layout and styling and it now plays audio locally (#5)
This reorganizes the extension to look a lot better, including the start of some reusable components, like the emoji buttons. It also makes the audio play locally at the same time as the injected audio. There's also a bunch of fun features like applying a border while playing and then removing it when it's done. All the colors and formatting and stuff is temporary until I get the final designs from Sharon.